refactor: create ColumnName class
